Address PR review: simplify casts in predicate helpers
- ExistsP: drop the redundant PBiPredicate cast; keep only the (V) cast that
  disambiguates the new P(PBiPredicate, V) vs P(PBiPredicate, GValue<V>) ctors.
- PredicatesUtil / UniGraphOrderStrategy: use Long.intValue() instead of
  (int)(long) for the boxed RangeGlobalStep.getHighRange() narrowing.

+++ unipop-core/src/org/unipop/process/order/UniGraphOrderStrategy.java
@@ -52,7 +52,7 @@ public class UniGraphOrderStrategy extends AbstractTraversalStrategy<TraversalSt
                 if (nextStep instanceof RangeGlobalStep){
                     if (step instanceof ReceivesPredicatesHolder) {
                         RangeGlobalStep rangeGlobalStep = (RangeGlobalStep) nextStep;
-                        int limit = rangeGlobalStep.getHighRange() > Integer.MAX_VALUE ? -1 : (int)(long) rangeGlobalStep.getHighRange();
+                        int limit = rangeGlobalStep.getHighRange() > Integer.MAX_VALUE ? -1 : rangeGlobalStep.getHighRange().intValue();
                         ((ReceivesPredicatesHolder) step).setLimit(limit);
                     }
                 }
